+++ src/main/java/lotto/pipeline/stage/AmountInputStage.java
@@ -1,4 +1,4 @@
-package lotto.controller.stage;
+package lotto.pipeline.stage;
 
 import lotto.domain.LottoRules;
 import lotto.exception.ErrorMessage;
