+++ src/main/java/lotto/pipeline/stage/Stage.java
@@ -1,4 +1,4 @@
-package lotto.controller.stage;
+package lotto.pipeline.stage;
 
 @FunctionalInterface
 public interface Stage<I, O> {
