+++ src/test/java/lotto/pipeline/stage/AmountInputStageTest.java
@@ -1,4 +1,4 @@
-package lotto.controller.stage;
+package lotto.pipeline.stage;
 
 import lotto.view.FakeInputView;
 import org.junit.jupiter.api.Test;
